Support for "long" domain names
nginx has limits on the size of its hash table entries, and long domains can
reach the limit. Failure has been observed with 40-character domains.

The failure manifests as a log message, and a failure to start nginx:

```
could not build server_names_hash, you should increase server_names_hash_bucket_size: 64
```

We can increase the limit, but doing so for all deployments may harm
performance. So we configure the larger limit only if it might be necessary.
As I'm currently unsure of the exact calculation needed to determine the exact
size, I chose to use the higher limit at domains over 35 characters in length.